Move business logic about whether an item can be requested to a separate class
