+++ app/components/item_request_link_component.rb
@@ -19,9 +19,7 @@ class ItemRequestLinkComponent < ViewComponent::Base
   end
 
   def render?
-    return false if aeon_pageable? || in_mediated_pageable_location? || in_nonrequestable_location? || @item.on_reserve?
-
-    current_location_is_always_requestable?
+    @item.request_link_policy.display?
   end
 
   private
@@ -43,31 +41,5 @@ class ItemRequestLinkComponent < ViewComponent::Base
       locale_key,
       default: :'searchworks.request_link.default'
     )
-  end
-
-  def current_location
-    @item.current_location.code
-  end
-
-  def current_location_is_always_requestable?
-    return true if current_location.end_with?('-LOAN') && current_location != "SEE-LOAN"
-
-    (Settings.requestable_current_locations[library] || Settings.requestable_current_locations.default).include?(current_location) ||
-      (Settings.unavailable_current_locations[library] || Settings.unavailable_current_locations.default).include?(current_location)
-  end
-
-  def in_mediated_pageable_location?
-    Settings.mediated_locations[library] == '*' ||
-      Settings.mediated_locations[library]&.include?(home_location)
-  end
-
-  def aeon_pageable?
-    Settings.aeon_locations[library] == '*' ||
-      Settings.aeon_locations.dig(library, home_location) == "*" ||
-      Settings.aeon_locations.dig(library, home_location)&.include?(@item.type)
-  end
-
-  def in_nonrequestable_location?
-    (Settings.nonrequestable_locations[library] || Settings.nonrequestable_locations.default).include?(home_location)
   end
 end
