+++ app/views/catalog/_index_location.html.erb
@@ -56,7 +56,7 @@
           <% location.items.each do |item| %>
             <tr>
               <td><%= item.callnumber %></td>
-              <td data-live-lookup-id="<%= document[:id] %>" data-status-target=".availability-icon" data-barcode="<%= item.barcode if item.live_status? %>" <%= "data-request-url='#{request_url(document, library: item.library, location: item.home_location, barcode: item.barcode)}'".html_safe if !location_request_link.render? && item.circulates? && !item.request_link.render? %>>
+              <td data-live-lookup-id="<%= document[:id] %>" data-status-target=".availability-icon" data-barcode="<%= item.barcode if item.live_status? %>" <%= "data-request-url='#{request_url(document, library: item.library, location: item.home_location, barcode: item.barcode)}'".html_safe if !location_request_link.render? && item.circulates? && !item.request_link_policy.display? %>>
                 <span class="availability-icon-wrapper">
                   <i class="availability-icon <%= item.status.availability_class %>"></i>
                   <span data-available-text="<%= t('searchworks.availability.available') %>" data-unavailable-text="<%= t('searchworks.availability.unavailable')  %>" class='status-text'>
@@ -72,7 +72,7 @@
                   <%= item.loan_period %>
                 <% end %>
                 <span class="request-link">
-                  <%= render item.request_link %>
+                  <%= render ItemRequestLinkComponent.new(item: item) %>
                 </span>
               </td>
             </tr>
